long long long

+++ main.c
@@ -7,11 +7,11 @@
 
 
 
-void kreduce(int * leastk, int * myids, unsigned long long int * myvals, int k, int world_size, int my_rank){
+void kreduce(int * leastk, int * myids, unsigned long int * myvals, int k, int world_size, int my_rank){
 	if(my_rank == 0){
 		int i;
 		int j;
-		unsigned long long int **leastkvals = (unsigned long long int**)malloc(world_size * sizeof(unsigned long long int*));
+		unsigned long int **leastkvals = (unsigned long int**)malloc(world_size * sizeof(unsigned long int*));
 		int **leastkids = (int**)malloc(world_size * sizeof(int*));
 		// printf("K is %d\n", k);
 		// for(i = 0; i < k * world_size; i++){
@@ -19,7 +19,7 @@ void kreduce(int * leastk, int * myids, unsigned long long int * myvals, int k, 
 		// }
 
 		for(i = 0; i < world_size; i++){
-			leastkvals[i] = (unsigned long long int*)malloc(k * sizeof(unsigned long long int));
+			leastkvals[i] = (unsigned long int*)malloc(k * sizeof(unsigned long int));
 			leastkids[i] = (int*)malloc(k * sizeof(int));
 		}
 		for(i = 1; i < world_size; i++){
@@ -39,7 +39,7 @@ void kreduce(int * leastk, int * myids, unsigned long long int * myvals, int k, 
 			// }
 		// }
 
-		unsigned long long int* leastval = (unsigned long long int*)malloc(k * sizeof(unsigned long long int));
+		unsigned long int* leastval = (unsigned long int*)malloc(k * sizeof(unsigned long int));
 		int* leastid  = (int*)malloc(k * sizeof(int));
 
 		int change = 1;
@@ -121,7 +121,7 @@ int main(int argc, char** argv){
 			struct node* ptr = readfile(document_path, query, dictionary_size, &list_size);
 			int i;
 			
-			unsigned long long int *myvals = (unsigned long long int*)malloc(list_size * sizeof(unsigned long long int));
+			unsigned long int *myvals = (unsigned long int*)malloc(list_size * sizeof(unsigned long int));
 			int *myids = (int*)malloc(list_size * sizeof(int)); 
 			// printf("List size is %d\n", list_size);
 			for(i = 0; i < list_size; i++){
@@ -144,7 +144,7 @@ int main(int argc, char** argv){
 
 		}
 		else{
-			unsigned long long int *myvals = (unsigned long long int*)malloc(k * sizeof(unsigned long long int));
+			unsigned long int *myvals = (unsigned long int*)malloc(k * sizeof(unsigned long int));
 			int *myids = (int*)malloc(k * sizeof(int));
 			// no memory allocation for leastk, slave processors will not be using that
 			int *leastk;
